Initial Commit to Master
This is the baseline code we received from ASUCI. The only change I
made was moving files around in the Project Navigator to be more
organized. All code files were in the top level directory in Xcode
basically, so I made folder groups and organized it a bit. No code was
touched though.