Add Google Analytics Pod, remove bundled lib
Transferring to using Pods for Google Analytics, as the project
previously decided to bundle the header files and static libraries with
it. It didn’t have the bundled libraries, so I’m just transitioning to
Pods for portability.

If you don’t have Pods: install Homebrew, use brew to install Pods, and
run `pod install` in the project directory. Email me for help.