Fixed the workings of pulling out the side menu
Now instead of having a hidden subview for the dragging, it works on
the map itself. If the map fails to get the pan gesture (within the
leftmost 30 points), then the reveal pan can work. Finally got the
right semantics for this.

Also if the front view is off to the side, the pan works for the entire
view to drag it back and close the rear controller.