Getting the MapView ready to handle the underlying data for multiple routes
It’s given all the routes via the side menu view (which I might change
to a singleton for the whole app later). From there it loads route
definitions for those that it doesn’t have, in case it already has the
information. It has everything in nice dictionaries for later use.