Add route stops and check boxes for side menu.
Added simple pins for the route stops, and uses iOS9 feature to set a
custom color. This does not work on iOS8, but it’ll be changed later
with a better fix than just a single color. See Facebook chat.

Also added a quick image for the checkboxes when selecting lines and
got the underlying logic done.