Temporary fix for a threading bug.
When routes are selected, the data model puts in a request for new
vehicle data at the same time as the map tries to read the data
currently in place. On my iPhone, it works fine but for the old iPad
Mini it crashes consistently. Need to find a safe and efficient
threading mechanism to solve this instead of just disabling the map’s
action.